r/aws_backup_vault_policy: Reduce visibility.
